Add CDNServeMux helper for origin server
ServeMux with an updateable handler, so that tests can pass their own
functions to make assertions about the requests received or emit their own
responses.

Includes a test to make sure that it works as intended. This is all in a
separate file so as not to clutter the main tests. It gets imported
automatically because it's part of the "main" package.

I had originally envisaged that we'd spawn a new server for each test, but
that's not practical because:

- DefaultServeMux doesn't allow you to unregister handlers and I didn't have
  any luck replacing the entire ServeMux associated with an http.Server
- We need to keep the server up and responding to probe health checks for as
  long as possible otherwise the CDN will mark it as down.