Panic on http.ListenAndServe error
So that we don't attempt to run the tests against another instance of `go
test` running on the same machine.

This won't catch, at least on OS X, the case of another process (no our
tests) binding to the TCP4 port only. It seems that Go will happily just
bind to TCP6.

Using the somewhat frowned panic() because we don't have access to testing.T
in init(). Part of the reason it's frowned upon is that we can't write a
test for it.