Mock CDN in a Vagrant VM using Nginx/Varnish
I found this useful to write and debug new tests while Mike was running
tests against the real service. Some of the behaviour we want to test is
just standard Varnish behaviour. We can mimic some other features with
minimal config changes.

The variable `$::ipaddress_vmhost` contains the IP address of the VM's host
on the other end of the private network - where `go test` would be running.

We'll need to be a careful about breaking this functionality when we
introduced SSL for backends. I'll leave a note in the issue.