Implement test for PURGE returning 403
This test, rather excitingly, proves that the behaviour has changed since we
implemented it in: https://www.pivotaltracker.com/story/show/62136616

The CDN now returns 200 and the body of a normal GET request. I suspect that
it's something to do with the `req.request = GET` hack that they suggested
implementing.

We should merge this but separately fix the config to make this test pass.