Move helper tests to helpers_test.go
So that they are logically separated from the actual helper code and match
the normal pattern of `foo_test.go` which tests `foo.go`.

I'm also removing the wrapper which called these individually - now that
they are in a `_test.go` file they will get picked up correctly themselves.
They can still be opted in with `-run Helper`. Plus we'll get the correct
errors if just one of them fail.