Mock for Cache-Control: private
This isn't default Varnish behaviour per:

https://www.varnish-cache.org/trac/ticket/477

However Fastly have implemented it in their standard config and I'm led to
believe that CloudFlare do the same, so we'll keep it in and mock it for
now. If it turns out that CloudFlare don't do it then we should consider
moving the tests away from the "standard caching behaviour" category.