Adjust Age Header test based on observation
We initially thought that the Edge would use the origin Age header to
effectively start the counter for its rewritten Age header - that an
Origin Age of 1000 would result in an Edge Age of 1002, after caching
for two seconds.

These tests have highlighted that this is not the case. Edge starts
the Age counter from 0 when it first caches.