Fix Age header test to reflect Varnish behaviour
We have made an interesting discovery during the writing of this
test.

Varnish appears to have slightly different behaviour with respect to
the Age header - at least between the Fastly version and our 3.0.5
mock test deployment.

In Varnish 3.0.5, the Age header would be increased based on the
value provided by origin -- that is to say an origin Age of 100 would
appear to be Age 100 when it went into the cache, and 105 5s later.

In Fastly Varnish, we were seeing this Age header get reset to 0, but
only in this test. In production, the former behaviour was noted.

@dcarley discovered that if the Cache-Control header was also set with
a max-age value set, the production behaviour was replicated.

This part of the Varnish source appears to relate to this functionality:

https://github.com/varnish/Varnish-Cache/blob/varnish-3.0.5/bin/varnishd/rfc2616.c#L42