testThreeRequestsAreCached expose complete resp
Expose the complete responseWriter in the callback from the helper because
TestCache404Response is going to need to call WriteHeader() in order to
create a 404 response.

The callback still occurs deliberately before writing the response body
because Write() will call WriteHeader(http.StatusOK) if WriteHeader() has
not already been called.