Implement TestCacheCacheControlMaxAge w/helper
Modify the testThreeRequestsAreCached helper so that it can optionally test
that a cached response expires after a given period of time.

The time sensitive nature of this test means that it is likely to fail if
the request/response latency is too high. It seems that we could do
something clever with time.Timer to test/mitigate that, but I don't think
it's worth the complexity.

It has now been renamed testRequestsCachedIndefinite and there is a
convenience function called testRequestsCachedIndefinite, used by existing
tests, which always passes a duration of zero.