Delete TestCacheDefaultTTL
This is going to be difficult to test for two reasons:

1. The value is going to be tightly coupled to the service configuration and
would likely need exposing as a CLI flag so that it can be modified.
2. It's going to be at least 60s and maybe as high as 30m, which is going to
make the tests run very slowly.

So I'm removing it and modifying the comment for TestCacheFirstResponse to
describe that it essentially covers the same case (but without the specific
cache period).