Add and implement TestCacheExpiresAndMaxAge
RFC2616 states that `Cache-Control: max-age=n` will take precendence over
`Expires: n` if both are present:

http://www.w3.org/Protocols/rfc2616/rfc2616-sec14.html#sec14.9.3

So test this by providing an `Expires` duration that is twice as long as
`Cache-Control`. We don't test the scenario described in the RFC, whereby
Expires is more restrictive, because we'd need to heavily modify our helper
to do that.