Implement TestFailoverOrigin5xxUseFirstMirror
Make a request for a new (uncached) URL, make the first server return a 503
response with an identifiable body, second server returns 200 with an
identifiable body, and third server raises an error if it's hit. An error is
raised if any server receives more than one request. We then verify the
status code and body of the response.