Start backend listeners in order, checking each
Introducing the StartBackendsInOrder func, which replaces our
previous confirmEdgeIsHealthy process.

Now, we start the backends 'from the bottom up', checking that
each backend is operational before moving onto the next higher priority
backend, up to and including the origin.

This ensures that we know the Edge has marked each backend as available,
giving us a consistent baseline to test against.