Modify 5xxServerStale time related variables
Renamed to indicate what they are doing - response TTL, response TTL with
buffer, and wait for saintmode. I've relocated the comment about saintmode
to the place where we're setting the var.

I've also changed the sleep on the second request (respTTLWithBuffer) to be
five times the value of the cache duration (respTTL). So that it will scale
accordingly if we change the TTL and it's more obvious why we're waiting.