Test mux.Stop() behaviour; inc additional helper code
In the process of writing this test, noted that the connection we are
making to the local listener (rather than via the CDN) was naturally
not affected by the 'close connection' request internal to the Stop()
method.

I have as a result added code to also process a close connection request
on any local client handler. I felt this was best placed in the Stop() logic
itself, rather than in the test, since it potentially will be necessary for
other local-only tests.