Make two additional requests when testing stale
To verify that we really are getting the stale object.

This is in part because the behaviour from the real CDN is currently wrong -
it seems that saintmode isn't being respected for the second request and
that it goes to origin twice and then gets a response from backup1.

I've moved the initialisation of expectedBody outside of the loop because we
now need to preserve the value between case statements.