Add confirmation that we served the close request
If a higher-level server is left running, the CDN will send
the request to that rather than our server. Hence, our server
will not send a Connection: close, and the whole backend system
is in an inconsistent state.

This ensures that if we do not receive the expected request, we
bail out with an error indicating that we must stop higher level
servers first.