Use httptest HTTP server - better client handling
We have a need to stop the HTTP server, so that we can
test the behaviour when backends are down. Previous work
using net/http have highlighted that we need to close off
client connections, as Keep-Alive is a factor.

A previous PR -- https://github.com/alphagov/cdn-acceptance-tests/pull/48
tackled this via sending a 'Connection: close' header, but in practise this
has opened up another set of problems and does not reliably work with
'multiple CDN nodes'

As a result, this commit starts the ball rolling again, but using the
httptest HTTP server, which records client connections and provides
methods to close them off.