Check body in TestFailoverErrorPageAllServersDown
Check the response body to see if we get the expected error page. At the
moment both the mocks and the real CDN give us Varnish's default error page.
Per the comment, this may change in the future.

Refactored to use the constant for the status code while I'm here.