Refactor out StoppableHttpListenAndServe
There's not much benefit to this being a separate function outside of
CDNBackendServer. It never returned a non-nil error so the anonymous
function goroutine in Start() was made redundant.

Now that we're using httptest.NewUnstartedServer() we have to create our
listener so we can catch errors from already bound ports there and it's
Start() method takes care of running Serve() in a goroutine so we don't have
to.