Create an ordered slice of backends
This describes their hierarchy as used by the CDN. It moves the initial
binding, with their names and ports, to init(). Then StartBackendsInOrder()
loops over them in reverse order starting and checking each one.

This will allow us to make the number of backends configurable for #51. In
addition to being able to resetting them in the correct order when we come
to check whether they are started for #50.