Add a IsStarted() method to CDNBackendServer
Which returns a boolean indicating whether the backend server has been
started or not. This will be used when resetting all backends.

I had hoped that net.Listener could tell us when the listener had been
Close()ed but it seems not. Another option was connecting to the address
reported by the server.Listener.Addr(), but that's quite a bit of work.

So instead I've opted to reset the server field back to nil after we've
closed it and report whether or not the field is bound.