CDNBackendServer.Start() and .ResetHandler()
Move the starting of a CDNBackendServer to the struct itself so that we can
lazily create them with the name and port, then start them later on. This
adds symmetry to the Stop() method and will come in useful when we want to
re-start an already bound but stopped instance.

Also created a ResetHandler() method, extracted from StartServer(), which
we'll be able to call to reset an already started backend back to a sane
state before each test.