Change description of Age header test
The expected behaviour for this changed during development when we found out
that the value in the Age header from origin will be incremented by edge if
the response is cached (has a CC header in this case).

Also moving the declaration of expectedAgeInSeconds to a constant at the top
of the test to make it look less magic.