Formatting errors in TestHelpersCDNServeStop
These were using Error instead of Errorf:

    helpers_test.go:58: possible formatting directive in Error call
    helpers_test.go:78: possible formatting directive in Error call

But in changing them to Errorf I also found that a boolean can't be printed
as `%q`. It should use `%t` instead.

I've also moved the expected value to a variable so that we don't repeat it
and it makes it more obvious why we're doing a comparison instead of just
`started` or `!started`.