[#60] Verify we're not purging in PURGE test
Improve the PURGE test to verify that we haven't actually purged the object
from edge's cache in addition to receiving the correct response code.

Because that's the actual behaviour we want to protect against and it's
quite conceivable that we could return a response code but still perform the
action.