[#28] Skipped test for `Vary: *`
Varnish didn't always support this. Added in:
varnish/Varnish-Cache@0d314640c075e0ff5a644def18cfe5792d22e107

However Fastly doesn't, I'm not sure how many other people do, and I'm also
not sure how important it is. So I'm going to add it for completeness with
the RFC but skip it so that it doesn't fail our tests for the moment.