Support custom TLS certificates
Allow a custom TLS certificate and private key to be passed in as CLI flags.
This will override the httptest.Server default self-signed certificate. It
will come in useful for CDN providers that perform strict checking of origin
certs.

This made me wonder whether we should write some tests to check whether
providers perform strict checking of cert chains and subject names. But
that's out of scope for now.