Remove disableBackendTLS and use stunnel in mocks
Supporting both HTTP and HTTPS backends is going to prove a lot of trouble.
The bug 998d01d was missed because I wasn't able to run all of the tests
locally.

So use stunnel to re-encrypt the request back from Varnish to our test
backend. This should work more-or-less the same as Fastly's TLS support in
Varnish.