Add TestCacheUniqueQueryParams
Per the test comment. This should never happen in Fastly, but Cloudflare
(and Akamai IIRC) have cache configuration level called "basic" that will
disregard the query param as a cache key.

Technically this should be picked up by all of our tests because the helper
uses the same path but different query params for each test. But it's better
to be explicit about it, especially if we change the helper in future. Hence
the additional assertions about req.URL.Path.