Add TestCacheUniqueCaseSensitive
Per the test comment. This was a feature that we once enabled in Akamai and
subsequently forgot about; it lead to some very non-deterministic behaviour
and took us a while to pluck up the courage to turn it off. We should
safeguard against this happening again in the future.

I went down a bit of a rabbit-warren trying to figure out how to deep-copy
*http.Request and *url.URL objects with pointer deferences. It turned out to
be much easier to just create two requests and set their respective Path and
RawQuery fields. The first two assertions check that this has actually
worked. The other parts of the request shouldn't matter.