Fix log message when backend starts with port 0
A port value of 0 can be used to let the OS pick an available high-numbered
port. This is handy and we use it in some of our helper tests. But it
results in a strange log message, which @mattbostock asked about:

    2014/08/01 12:42:37 Started server on port 0
    2014/08/01 12:42:37 Started server on port 0
    2014/08/01 12:42:37 Started server on port 0

Fix this by parsing the assigned port and copying it to s.Port so that the
log line does the right thing again. This now looks like:

    2014/08/05 15:58:59 Started server on port 57552
    2014/08/05 15:58:59 Started server on port 57552
    2014/08/05 15:58:59 Started server on port 57555
    2014/08/05 15:58:59 Started server on port 57557
    2014/08/05 15:58:59 Started server on port 57560