[#75802914] Allow whitespace in X-Forwarded-For
The draft specification, which is probably the best we'll ever have, says
that whitespace between values is optional:

- http://tools.ietf.org/html/draft-ietf-appsawg-http-forwarded-10#section-7.1

Although it's quite common for there to be spaces, CloudFlare and some other
reverse-proxies don't add them, and it's even more common for software that
reads the header to do a split+trim.

So relax this test by checking the total number of values and then the value
of each after stripping whitespace.