Fail test when no vendor implementation known
Fail, rather than skip, a test if we haven't written test logic specific
to that vendor.

The aim of this is that new vendors will causes the tests to fail,
meaning that either:

a) the test should be modified to skip that test if the feature is
_known_ not be supported by the vendor, or,

b) the test should be modified to support the vendor's unique
implementation of the feature if it's supported, or,

c) the vendor CDN configuration needs modification to make the tests
pass

This will allow us to distinguish between features we know to be
unsupported, features that are supported but we don't yet test correctly
for, and features which have not yet been correctly configured in our
CDN account.

Note that we specify Fastly as the vendor when running the mock tests
(also used by Travis) as Fastly are our current vendor and the mock
virtual machine roughly emulates Fastly's service. If we didn't specify
Fastly when running the Travis CI tests, the tests would fail with
'notImplementedForVendor' errors.

* * *

Also, fix a syntax error where `t.Skip()` was called as `t.skip()`.