Add 'Guru Mediation' error page to mock config
Add a 'Guru Mediation' (sic) error page to the Varnish mock
configuration to match Fastly's error page which contains this same
typo.

Since the Varnish mock configuration already mimics Fastly's behaviour,
it makes sense to extend this to the error page. We can then look at
mimicking other CDN vendors using the mocks as necessary in future.

This page is adapted from Varnish's built-in `default.vcl`[1] with a
noticeable difference that I had to add `+` as a concatenation error,
otherwise the VCL fails to compile with the error:

    Notice: /Stage[main]/Varnish/Exec[varnish-restart]/returns: Expected ';' got 'obj.status'
    Notice: /Stage[main]/Varnish/Exec[varnish-restart]/returns: (program line 174), at
    Notice: /Stage[main]/Varnish/Exec[varnish-restart]/returns: ('input' Line 157 Pos 15)
    Notice: /Stage[main]/Varnish/Exec[varnish-restart]/returns:     <title>"} obj.status " " obj.response {"</title>
    Notice: /Stage[main]/Varnish/Exec[varnish-restart]/returns: --------------##########----------------------------
    Notice: /Stage[main]/Varnish/Exec[varnish-restart]/returns:
    Notice: /Stage[main]/Varnish/Exec[varnish-restart]/returns: Running VCC-compiler failed, exit 1
    Notice: /Stage[main]/Varnish/Exec[varnish-restart]/returns:
    Notice: /Stage[main]/Varnish/Exec[varnish-restart]/returns: VCL compilation failed

This is also necessary to make the CI tests pass since we're modelling
our Varnish behaviour on Fastly.

[1]:
https://www.varnish-cache.org/trac/browser/bin/varnishd/default.vcl?rev=3.0