Move Travis CI make task into a bash script
Move the make task for Travis CI to a bash script. This is in
preparation for removing the Makefile.

Also, simplify the command-line options for `go test`; `-test.v` has the
same behaviour as the shorter `-v` and `-run` is removed as all tests
are run by default so this was unnecessary.