Skip 'install' step
Tell Travis CI to run `true(1)` for the `install` step.

This is to prevent Travis from running `go get -v ./...`, which fails[1]
with:

    ./helpers.go:117: undefined: edgeHost
    ./helpers.go:146: undefined: client
    ./helpers.go:147: undefined: requestSlowThreshold
    ./helpers.go:150: undefined: debugResp
    ./helpers.go:215: undefined: client
    ./helpers.go:276: undefined: edgeHost
    ./helpers.go:279: undefined: originServer
    ./helpers.go:346: undefined: originServer

I believe this is due to the way we're using flags in a suite of tests
(running `go test` works fine).

We don't need to run `go get` as we don't have any package dependencies
outside of Go's standard library. By skipping Travis' `install` step, we
prevent it from running `go get`.

Note that it that this wasn't a problem before because Travis apparently
doesn't run `go get` if it detects a Makefile[2], which is consistent
with the documentation here:
http://docs.travis-ci.com/user/languages/go/#Default-Test-Script

[1]: https://travis-ci.org/alphagov/cdn-acceptance-tests/builds/33156531
[2]: https://travis-ci.org/alphagov/cdn-acceptance-tests/builds/33047132