Support Cloudflare's 'EXPIRED' status
Cloudflare's `CF-Cache-Status` header may contain one of the
following[1]:
  - HIT
  - MISS
  - EXPIRED
  - STALE
  - IGNORED
  - REVALIDATED

We already support `HIT` and `MISS`; this commit adds support for
`EXPIRED`, which is used when a previously-requested object is requested
one further time after its TTL. We specify the object's TTL using the
`Cache-Control` `max-age` header.

I haven't implemented `STALE`, `IGNORED` or `REVALIDATED` because they
only occur under very specific circumstances and so are less useful to
us.

Note that I expected to see `IGNORE` rather than `MISS` on an object's
first request, as per Cloudflare's documentation, but in reality the
`IGNORE` status did not occur in my testing.

[1]:
https://support.cloudflare.com/hc/en-us/articles/200168266-What-do-the-various-CloudFlare-cache-responses-HIT-Expired-etc-mean-