Pass http.Request to testRequestsCached*
I'm about to move a test from TestNoCache to TestCache that depends on
request headers being set (rather than response headers). So we need a way
to pass http.Request objects in to the helper. This adds one LoC to each
TestCache but does at least make it look more like TestNoCache.