Remove checkForSkipFailover() for TestServeStale*
Remove the `checkForSkipFailover()` check, which prevents a test from being
run if we don't have multiple backends, so that we can run the ServeStale
tests against CloudFlare who have implemented this but (not yet) multiple
backends.

These tests are now agnostic about how many backends exist in addition to
origin. If any exist, then we set their handler to assert a test error when
they receive a request.

This has surfaced the quirk that we were already aware of with our Varnish
serve-stale implementation. The test for OriginDown works if you run it with
`-skipFailover` because Varnish never successfully health checks the
additional backends and always considers them sick. But if you run it with
multiple backends then it doesn't serve stale on the first restart and hits
the first backup server on the second restart. I'll look at either fixing
this or temporarily skipping the test for that combination of arguments.