Merge the two ServeStaleOriginDown tests
I split these into two tests in 7cce799 because we found that the behaviour
was different depending on whether you wait for the health check to expire
or not. The intention was to stop the tests failing and document the nuances
of the current behaviour.

In hindsight I think this was a mistake. They are confusing to read and make
it unclear what our *desired* behaviour is. If we're unable to get the right
implementation but we're confident about the test implementation then we
should use `t.Skip(notSupportedByVendor)` in the meantime, as we have done
for some other tests recently.

So merge them back into a single test for serving stale when origin is down.
We use `respTTLWithBuffer`, like the other stale test, to ensure that we're
well beyond the object's normal TTL. We do this before stopping the origin
to ensure that we're not relying on health check expiry.

The one difference now between the OriginDown and Origin5xx tests are that
the latter also checks that the stale object isn't used when origin comes
back. We might want to implement this for OriginDown later. But I suspect
that it may be complicated by removing `checkForSkipFailover()` from these
tests - which I'll explain in a following commit.

NB: The old test relied on `healthCheckExpire` to also ensure that we were
beyond the TTL. This was a mistake - it just happened that the expire time
was longer than the TTL.