Fix serve stale in mocks by using fake backend
Fix `TestServeStaleOriginDown` when run against the mocks without
`-skipFailover`. This was failing - for the same reason that we split the
tests in 7cce799 - when origin is down and the health check hasn't noticed
yet then Varnish would:

- reattempt the request against origin on the first restart
- attempt the request against the backup server on the second restart
- invalidate the stale object with the response from the backup server

We can workaround this by using the "evil hack" described here:

- https://www.varnish-software.com/static/book/Saving_a_request.html#example-evil-backend-hack

It creates a new backend which will always be "sick"; never receives a real
request. It uses a port which shouldn't be bound on the guest, a request
that isn't valid HTTP, an interval that means it will rarely be polled, and
an initial value which means that it's immediately sick when Varnish first
starts.

This causes Varnish to use the stale/grace object (if it exists) on the
first restart and never invalidate it with a response from a backup server.
This now works the same regardless of whether `skipFailover` is or isn't
used.