Document Cache-Control HTTP request behaviour
We test the behaviour of the CDN edge for `Cache-Control` headers in
HTTP responses but not for `Cache-Control` _request_ headers.

RFC 7234 section 5.2 states that a cache `MUST` obey the requirements of
the `Cache-Control` directives defined in section 5.2, which includes
these `Cache-Control` request headers (defined in section 5.2.1).

Note that none of Cloudflare, Fastly or our Varnish mock configuration
respects above RFC; this commit documents the actual behaviour here for
reference.