Use mime.TypeByExtension for fixture Content-Type
DRYs up the tests significantly. This doesn't suffer from the same problems
as http.DetectContentType described in 40048ca whereby it would detect JS
and CSS as generic `text/plain`.

The builtin map works sufficiently on OS X which doesn't have
`/etc/mime.types` to supplement it. Although, just to be safe, I have added
an additional check to make sure that we don't get an empty or generic type.

Thanks to @mattbostock for indirectly finding this.