Set req.URL.Path in testResponseNotManipulated
To give the CDN some additional help in determining the type. The
`Content-Type` header should be sufficient, but we can't get the image tests
to fail on CloudFlare when the right boxes are ticked and we've noticed that
they sometimes care more about file extensions than content.

This won't affect the random UUID that forms part of the request as a query
param because they exist in a different `req.URL` field.