Remove vendor switch for True-Client-IP
This is now implemented by both of our providers, so we no longer need to
switch on the vendor flag.

Include a comment in the test description that this needs to be consistent
across all vendors. Too long for the comment - but this is because if we
accepted two values and tested them in order, then it would be possible to
spoof one of them.