[#75802710] Test Age header when not set by origin
This compliments the existing TestRespHeaderAgeFromOrigin test and will help
us clarify how much CloudFlare have implemented.

The first value of "0" passes for Fastly and Varnish, but fails for
CloudFlare who return no header on the first response. The RFC isn't clear
which is the right behaviour but I'm leaning towards the former.