Change expected body text to match new error page
We want to display a custom synthetic HTML page from the CDN when
the origin and both sets of mirrors are down, rather than the
standard "Guru Mediation/Meditation" page.

This commit pulls in some text from that page that we can assert
with and updates the mock Varnish config to reflect the change.