Consider all HEAD requests to be probes
Consider all HTTP HEAD requests to be backend probes.

We treat backend healthcheck probes (originating from the CDN)
differently to normal requests as they could skew our test results.

Any HEAD request for the `/` path was considered as a probe.

When testing a new VCL configuration, we needed to use a probe endpoint
other than `/` (that endpoint returns a 404 for this particular
service).

So relax the check such that any `HEAD` HTTP requests are considered to
be probes. We don't have any acceptance tests specific to HEAD requests
currently.

Also, update the README to highlight this code.